feat: 字体配置出码优化
