+++ docs/daily-note.md
@@ -541,13 +541,12 @@ TODO:
 - [x] 资源面板-模板列表
 
   - client
+
     - [x] 渲染出模板列表
-    
+
     - [x] 检索 选择？
-    
+
     - [x] 渲染逻辑？
-      
-      
 
 - [x] 资源发布
