+++ src/pages/editer/components/config/components/global-config/font-form/index.tsx
@@ -12,11 +12,9 @@ import {
 import { IconSave } from '@arco-design/web-react/icon'
 import { FontList } from '@/typings/database'
 import FontCascader from '@/components/custom-form-inner/font-cascader'
-import { FontFormData } from '@/typings/common/editer-config-data'
-import {
-  filterFontByType,
-  updateFontConfigToElement,
-} from '@/common/utils/font'
+import { filterFontByType } from '@/components/custom-form-inner/utils'
+import { FontFormField } from '@/typings/common/editer-config-data'
+import { updateFontConfigToDOM } from '@/common/utils/font'
 import { useGlobalConfig } from '@/common/hooks/editer-data'
 import HelpTip from '@/components/help-tip'
 
@@ -35,21 +33,21 @@ const { Item: ListItem } = List
 const { Meta: ListItemMeta } = ListItem
 const { Item: CollapseItem } = Collapse
 
-const FontForm: React.FC<FontFormProps> = ({ fontList }) => {
+const GlobalFontForm: React.FC<FontFormProps> = ({ fontList }) => {
   const { configData, updateGlobalConfig } = useGlobalConfig()
 
-  const [fontForm] = Form.useForm<FontFormData>()
+  const [fontForm] = Form.useForm<FontFormField>()
 
   const platFormFontList = React.useMemo(
     () => filterFontByType(fontList, 'platform_provide'),
     [fontList]
   )
 
-  const handleFormChangeHandler: FormProps<FontFormData>['onChange'] = (
+  const handleFormChangeHandler: FormProps<FontFormField>['onChange'] = (
     value,
     values
   ) => {
-    // 更新 的字段不是  globalFont
+    // 更新 的字段不是  globalFont 退出执行
     if (!value.globalFont) return
     const [, globalFont] = value.globalFont
     const usedFont = values.usedFont || []
@@ -63,7 +61,11 @@ const FontForm: React.FC<FontFormProps> = ({ fontList }) => {
   }
 
   const handleSaveClick = () => {
-    updateFontConfigToElement(fontForm.getFieldsValue(), fontList, true)
+    updateFontConfigToDOM(
+      fontForm.getFieldsValue().globalFont?.[1],
+      fontList,
+      true
+    )
 
     updateGlobalConfig({
       fontConfig: fontForm.getFieldsValue(),
@@ -135,4 +137,4 @@ const FontForm: React.FC<FontFormProps> = ({ fontList }) => {
   )
 }
 
-export default FontForm
+export default GlobalFontForm
